chore(api-tests): move sanitization test to review workflows test suite
